+++ R/mod-page-xforms.R
@@ -95,17 +95,17 @@ page_xforms_server <- function(id, data_name_in = NULL) {
 
   init_data_name <- NULL
   if (!is.null(data_name_in)) {
-    if (shiny::is.reactive(data_name_in)) {
-      init_data_name <- data_name_in
-    } else if (checkmate::test_data_frame(data_name_in)) {
-      init_data_name <- deparse(substitute(data_name_in))
-    } else if (checkmate::test_string(data_name_in)) {
-      if (exists(data_name_in, envir = .GlobalEnv)) {
+    try(silent = TRUE, {
+      if (shiny::is.reactive(data_name_in)) {
+        init_data_name <- data_name_in
+      } else if (checkmate::test_data_frame(data_name_in)) {
+        init_data_name <- deparse(substitute(data_name_in))
+      } else if (checkmate::test_string(data_name_in)) {
         if (checkmate::test_data_frame(get(data_name_in, envir = .GlobalEnv))) {
           init_data_name <- data_name_in
         }
       }
-    }
+    })
   }
 
   moduleServer(