+++ NEWS.md
@@ -1,6 +1,7 @@
 # Unreleased version
 
 - Transformations: fix bug where missing values in numeric columns were not treated as missing values
+- Transformations: when invoking the addin in RStudio, if a variable holding a data frame is selected then that data frame will be used
 - Transformations: allow wizard to be initialized with an existing data frame
 - Transformations: allow wizard to work with reactive data